+++ wp-content/plugins/core/src/Service_Providers/Theme_Provider.php
@@ -11,7 +11,7 @@ use Tribe\Project\Theme\Image_Links;
 use Tribe\Project\Theme\Image_Sizes;
 use Tribe\Project\Theme\Image_Wrap;
 use Tribe\Project\Theme\Nav\Nav_Attribute_Filters;
-use Tribe\Project\Theme\Oembed_Wrap;
+use Tribe\Project\Theme\Oembed_Filter;
 use Tribe\Project\Theme\Resources\Editor_Styles;
 use Tribe\Project\Theme\Resources\Emoji_Disabler;
 use Tribe\Project\Theme\Resources\Fonts;
@@ -43,8 +43,8 @@ class Theme_Provider implements ServiceProviderInterface {
 		$container[ 'theme.images.responsive_disabler' ] = function ( Container $container ) {
 			return new WP_Responsive_Image_Disabler();
 		};
-		$container[ 'theme.oembed.wrap' ] = function ( Container $container ) {
-			return new Oembed_Wrap();
+		$container[ 'theme.oembed' ] = function ( Container $container ) {
+			return new Oembed_Filter();
 		};
 		$container[ 'theme.supports' ] = function ( Container $container ) {
 			return new Supports();
@@ -93,7 +93,7 @@ class Theme_Provider implements ServiceProviderInterface {
 		$container[ 'service_loader' ]->enqueue( 'theme.images.wrap', 'hook' );
 		$container[ 'service_loader' ]->enqueue( 'theme.images.links', 'hook' );
 		$container[ 'service_loader' ]->enqueue( 'theme.images.responsive_disabler', 'hook' );
-		$container[ 'service_loader' ]->enqueue( 'theme.oembed.wrap', 'hook' );
+		$container[ 'service_loader' ]->enqueue( 'theme.oembed', 'hook' );
 		$container[ 'service_loader' ]->enqueue( 'theme.supports', 'hook' );
 		$container[ 'service_loader' ]->enqueue( 'theme.resources.login', 'hook' );
 		$container[ 'service_loader' ]->enqueue( 'theme.resources.legacy', 'hook' );
