+++ wp-content/themes/core/functions.php
@@ -17,10 +17,6 @@ if ( ! function_exists( 'core_theme_setup' ) ):
 
 function core_theme_setup() {
 
-	// @Jonathan
-	// Port Me To Core
-	include_once 'port-me-to-core/Embeds.php';
-
 	// Template Tags
 	include_once 'template-tags/titles.php';
 	include_once 'template-tags/images.php';
